+++ MatrixSDK/Utils/MXBase64Tools.h
@@ -34,11 +34,11 @@ NS_ASSUME_NONNULL_BEGIN
 
 #pragma mark - Data
 
-+ (NSData *)dataFromUnpaddedBase64:(NSString *)unpaddedBase64;
-+ (NSString *)unpaddedBase64FromData:(NSData *)data;
-
+// The base64 string can be padded or unpadded
 + (NSData *)dataFromBase64:(NSString *)base64;
+
 + (NSString *)base64FromData:(NSData *)data;
++ (NSString *)unpaddedBase64FromData:(NSData *)data;
 
 @end