+++ CHANGES.rst
@@ -8,7 +8,7 @@ Improvements:
  * 
 
 Bugfix:
- * 
+ * MXBase64Tools: Make sure the SDK decode padded and unpadded base64 strings like other platforms (vector-im/riot-ios/issues/3667).
 
 API Change:
  *